REFACTOR - change palce of file like FSD arcitecture
