+++ src/features/module/types/form.ts
@@ -1,7 +1,7 @@
 import { ValidationRule } from 'src/shared/ui/VForm/validation/rules';
 import { useText } from 'src/shared/ui/VForm/form-inputs';
 import { IInputString } from 'src/shared/ui/VForm/form-inputs/base-inputs';
-import { IModule } from '../types';
+import { IModule } from './module';
 
 export interface IModuleFormConfig {
   id: string;
