+++ src/pages/IndexPage.vue
@@ -1,37 +1,5 @@
 <template>
-  <q-page class="row items-center justify-evenly">
-    <example-component title="Example component" active :todos="todos" :meta="meta"></example-component>
-  </q-page>
+  <q-page class="row items-center justify-evenly"> INDEX PAGE </q-page>
 </template>
 
-<script setup lang="ts">
-  import { Todo, Meta } from 'components/models';
-  import ExampleComponent from 'components/ExampleComponent.vue';
-  import { ref } from 'vue';
-
-  const todos = ref<Todo[]>([
-    {
-      id: 1,
-      content: 'ct1'
-    },
-    {
-      id: 2,
-      content: 'ct2'
-    },
-    {
-      id: 3,
-      content: 'ct3'
-    },
-    {
-      id: 4,
-      content: 'ct4'
-    },
-    {
-      id: 5,
-      content: 'ct5'
-    }
-  ]);
-  const meta = ref<Meta>({
-    totalCount: 1200
-  });
-</script>
+<script setup lang="ts"></script>
